feat(tui): extract sleep inhibitor util and address PR feedback
