+++ codex-rs/utils/sleep-inhibitor/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "sleep-inhibitor",
+    crate_name = "codex_utils_sleep_inhibitor",
+)
