+++ codex-rs/utils/sleep-inhibitor/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "codex-utils-sleep-inhibitor"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[lints]
+workspace = true
+
+[target.'cfg(target_os = "macos")'.dependencies]
+core-foundation = "0.9"
+libc = { workspace = true }
+tracing = { workspace = true }
